Moved content serializers to a different folder
