+++ Distroir.CustomSDKLauncher.Core/Managers/ContentSerializers/BinaryFileSerializer.cs
@@ -15,14 +15,11 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Runtime.Serialization.Formatters.Binary;
-using System.Text;
 
-namespace Distroir.CustomSDKLauncher.Core.Managers.Serializers
+using System.IO;
+using System.Runtime.Serialization.Formatters.Binary;
+
+namespace Distroir.CustomSDKLauncher.Core.Managers.ContentSerializers
 {
     public class BinaryFileSerializer<T> : ContentSerializer<T>
     {
